feat: add ViewModel
