--- src/Life-Balance.WebApp/Models/RegisterViewModel.cs
@@ -1,22 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Life_Balance.WebApp.Models
-{
-    public class RegisterViewModel
-    {
-        [Required]
-        [Display(Name = "Email")]
-        public string Email { get; set; }
-
-        [Required]
-        [DataType(DataType.Password)]
-        [Display(Name = "Password")]
-        public string Password { get; set; }
-
-        [Required]
-        [Compare("Password", ErrorMessage = "Passwords do not match")]
-        [DataType(DataType.Password)]
-        [Display(Name = "Confirm password")]
-        public string PasswordConfirm { get; set; }
-    }
-}
