+++ src/Life-Balance.WebApp/ViewModels/ForgotPasswordViewModel.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Life_Balance.WebApp.ViewModels
+{
+    public class ForgotPasswordViewModel
+    {
+        /// <summary>
+        /// Email.
+        /// </summary>
+        [Required]
+        [EmailAddress]
+        public string Email { get; set; }
+    }
+}
